processors/refine-descriptor/udpipe-json/requirements.txt
ufal.udpipe==1.2.0.3
Unidecode==1.1.1
